requirements.txt
Flask
feedparser
APScheduler
SQLAlchemy
python-dateutil